refactor(cli): move the partition group into its own module
Cuts `gpio partition` out of cli/main.py into
geoparquet_io/cli/commands/partition.py, unchanged. Clean cut: the group owns
no helpers beyond its own commands.

`tests/test_cli_api_call_parity_scaffold.py` resolved `partition_by_h3_impl`
and `partition_by_quadkey_impl` through `geoparquet_io.cli.main` and now
passes `module=cli_partition`.

Eight pre-existing `arg-type` errors across the partition commands were
suppressed by the mypy per-module override on `geoparquet_io.cli.main`, so the
override entry follows the code rather than the errors being fixed here.

tests/data/cli_surface.json is byte-identical.
